Move SetFD to testdb
